requirements.txt
httpx
pywebview